[509912] Use the EValidatorAdapter for PropertiesPackage.INSTANCE
The validation of interpreted expression rely on the EMF.Validation
audits.

This adapter is needed to trigger the EMF.Validation audits when a
"plain EMF Core" validation is launched.  This happens for instance in
the VSM editor when using right-click->Validate on an element.

Bug: 509912
Change-Id: I94b2f7afebfcb7954869a9e5abd01d696ad881af
Cherry-picks: 0ae9613cadd1d8164666e6c4e813fb503920a760
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>