[509930] Make AQL bridge support Java typed variables
The property views support rely on a variable named 'input' which is
present in every interpreted expression. This variable is of type
org.eclipse.sirius.ui.properties.internal.SiriusInputDescriptor but even
if it is specified as such in the code the AQL integration is not
transforming this declaration into the corresponding AQL type and just
fall back to EObject.

This commit adapts TypeName and VariableType so that an instance of
VariableType can be created from a Java class and then declared as such
to AQL when validating the expression or calling content-assist.

Bug: 509930
Change-Id: I7e6e31af2149de85916de9226697b734e5a1d673
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>