[509930] enable Properties checks also for subtypes in another EPackage
Some specific widgets can be contributed by other plugins and will then
be in other EPackages. Those widget might inherit from the original one
and as such should be subject to the same validation rules. 
ExtReferenceDescription is an example of that situation. 
This commits makes sure that in a situation of the EObject's EClass is
from a different EPackage but inherits from an EClass in the Properties
epackage the checks are still enabled 

Bug: 509930
Change-Id: I902996e34615ac5d6b1f128cfee4cb64f272c157
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>