[501073] Auto-expand just enough to reveal the representations
An unintended side-effect of the previous commit for this ticket was
that when a VSM references the semantic metamodel(s) (which is
encouraged), those metamodels were also expanded, which is not
particularly useful. It also expanded the custom palettes if present.

Only expand the minimum required to reveal the representation (and
representation extensions).

Bug: 501073
Change-Id: I31089d79dbef195beda3cbccc0440b0953ed1954
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>