[501073] Auto-expand just enough to reveal the representations
An unintended side-effect of the previous commit for this ticket was
that when a VSM references the semantic metamodel(s) (which is
encouraged), those metamodels were also expanded, which is not
particularly useful. It also expanded the custom palettes if present.

Only expand the minimum required to reveal the representation (and
representation extensions).

This version of the patch is rewritten in Java 7, the original in
commit 5745590dcefe9de9d9f75b0f7cae27a9bcfabcca used Java 8-only
syntax and APIs.

Bug: 501073
Change-Id: Ia6243e87ffa7cdce34189a0ba4b488d78579c401
Cherry-picks: I31089d79dbef195beda3cbccc0440b0953ed1954
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>