[500253] Fix compilation pb during tests execution with Tycho surefire
This problem is explained/detailed in:
* https://bugs.eclipse.org/bugs/show_bug.cgi?id=343152
* https://wiki.eclipse.org/Tycho/Testing_with_Surefire
* https://bugs.eclipse.org/bugs/show_bug.cgi?id=343156
* https://bugs.eclipse.org/bugs/show_bug.cgi?id=386988

To resume, Tycho surefire executes tests in an Eclipse initialized with
OSGi but not with PDE. So the PDE target platform is not fully
initialized and all "tools" using it, like plugin compilation, fail.

This commit uses the solution [1] that loaded a new PDE target platform
created from the known OSGI bundles.
The new static method
org.eclipse.sirius.tests.support.api.TestsUtil.setTargetPlatform() must
be called for each test that need to compile a plug-in in the execution
runtime.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=343152#c9

Bug: 500253
Change-Id: I2385535f63a52929e918df44bf49c1f2037e3cd5
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>