[510126] Use ToolDescription precondition to infer variable types
Leverage the type inference in predicate expressions so that a tool with 
element.oclIsKindOf(ecore::EPackage) in a precondition expression will
have 'element' infered as an EPackage.

This is only enabled for the generic tools so far for which this is
critical as they are not referencing any Mapping and as such we don't
have a lot of type information. 

Enabling it for every subclass of AbstractToolDescription right now
might lead to infinite loops in the validation as detected by
VSMVariableTypesValidationTest.

Bug: 510126
Change-Id: Ib641ab2b0f55128977865294bbf2edc2f7341239
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>