[510126] Always use DSemanticDiagram instead of DDiagram
This commit changes the all the variable declarations in the ecore and
generated code to refer to DSemanticDiagram instead of DDiagram. It is
quite common for queries to try to access the semantic root of a
diagram, casting in those case seems useless as Sirius will always
manage DSemanticDiagram instances.

Bug: 510126
Change-Id: Ibe2b2e365ce1434da9966dd9a6f80fd518ef2863
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>