[510692] Introduce a WARNING if a typename is not accessible
This change leverage the fact that AQL will return errors if a type is
not accessible in the classpath. The validation rule will actually build
a very simple AQL expression checking the type, and if any error come up
then it will deduce the type is not accessible.

If AQL is not there, then no error is returned and as such the
validation rule will never issue a WARNING.

Bug: 510692
Change-Id: I4148ed130f0aa50decb99137abf8f5f0bf7b07b2
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>