[test] Ignore jdt error of kind "Resource folders already exists"
Sometimes, there is an error during test MigrationOnVsmEditorReloadTest.
This error is of kind "Resource
'/.org.eclipse.jdt.core.external.folders/.link0' already exists.". It is
possible to reproduce this problem by creating a VSP but only with a
runtime environment. It seems that this problem occurs when JDT tries to
create an external folder for the dependencies existing in the workspace
("org.eclipse.sirius/bin" and
"org.eclipse.sirius.common.acceleo.aql/bin"). Pending resolution of the
problem, this workaround avoids to have sometimes test with error.

Change-Id: I9d9a6cc83ecb9b27096fc0eede1b08cb3bea83e7
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>