[507056] Fix potential leak and lower the default target pixels number
After experimenting with the values of _Pref_DiagramExportSizeMax using
the ExportDiagramAsImagesTest. Setting the limit to 
* 100 Millions of pixels => the test fails quickly with an OOM
*  50 - heap usage goes up to 1.9Gb
*  30 - heap usage goes up to 1.6Gb
*  20 - heap usage goes up to 1.2Gb
*  10 - heap usage goes up to 1.2Gb

Seems like 10 millions of pixels would be a sweet spot as it brings a
more than noticeable improvement in resolution and have a reasonnable
cost but on some Linux platforms. On the other hand the tests results
are varying vastly depending on the OS, platform and/or versions of
underlying libraries like Cairo.

After experimenting with those different values it seems 10 millions of
pixels as a target -- by default -- leads to VM crash (double free()
calls in native code) which are likely to be related to specific
versions of GTK and Cairo (we have consistently seen this on ubuntu
machines but not on other distributions). 

After many tests it appears 4.125M is the maximum value we could find
which doesn't lead to those crashes. As such this commit lower the
target to 4.125M pixels and adapt the test as no case is supposed to
fail anymore.

Bug: 507056
Change-Id: Iaba6b641499bed6d8a9d48ea4ae8387e7bfa6829
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>