[510040] Add validation for viewpoints activation in session editor
Viewpoints with missing activated viewpoint dependencies now cannot be
activated. When trying to activate, an error message tells the user what
dependencies are missing and checkbox is marked with a black square
instead of a tick meaning it has not been activated. Selecting a
viewpoint with missing dependencies also triggers the display of the
error message.

Viewpoints activation now react to double click.

When deactivating a viewpoint, all activated viewpoints depending on it
are also deactivated.

Bug: 510040
Change-Id: I35acf91d05662f401fdc3ac275c320fc56383d9c
Signed-off-by: pguilet <pierre.guilet@obeo.fr>