[512443] EditModeDecorator should compute isBroken image first
The previous commit on EditModeDecorator optimized the decorator refresh
by removing duplicates in calls to permission authority and isBroken
checks (see shoulBeDecorated() and getDecorationImage()). 

The logic of the previous version of shouldBeDecorated was moved to
getDecorationImage but this changed the image priority: now the lock
status is displayed even if the edit part is broken whereas we want to
display only the red cross when the part is broken even if the lock
status could lead to a specific decoration.

This commit set back the previous getImage() computation priority.

Bug: 512443
Cherry-picked-from: 512292
Change-Id: Ib146180cf051f17095787563e5ec121406a5825c
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>