[430948] Use the default font, not just the default size for trees
The objective of the new preference is to make Sirius-defined trees
look as "native" as possible. Under Windows, the Sirius default font
very close to the default/system one, but this is not the case
everywhere. Make sure we actually use the system default font, not
just its size, when the preference is enabled.

Bug: 430948
Change-Id: I9f6b0f31ee71b8a19f2c40910fbcd6bc6aa48a0f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>