[507474] Make sure Sirius 5.0 is compatible with Guava 21
Eclipse Oxygen may switch to Guava 21 as the default (only?) version
in the SimRel, so we need to make sure we're compatible with it.

* Open the version ranges to include Guava 21.
* Remove all usages of Objects.firstNonNull(), which does not exist
  anymore in Guava 21.

Not that at this point, other dependencies of Sirius have different
version constraints, so it's not enough to ensure v21 is the one
resolved:
* org.eclipse.xtext.xbase.lib 2.11.0.v20170124-1424: [14.0.0,19.0.0)
* org.eclipse.xtext.util 2.11.0.v20170130-0808: [14.0.0,19.0.0)

Xtext considers moving to Guava 21 for v2.12, see
https://github.com/eclipse/xtext-lib/issues/30.

In the meantime, the build still resolves to Guava 15, but a test was
made by removing Xtext from the build, and Sirius builds fine when
forcing Guava 21 to be really used.

Bug: 507474
Change-Id: I178bcd70fe3fba95cda247473d1ed18f5206acc9
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>