[test] Try to make CustomizationPropertySectionsTests more reliable
The test fails quite often on CI servers, but not always and never
locally, which points to some timing issue.

The failure point (at least the one I observed) is always the same.
The test:
1. changes the value of one text field in the property section;
2. clicks on a button in another widget/section of the same element;
3. the button open a dialog whose contents depends on the state of the
   underlying model;
4. some assertion fails because the dialog does not display the
   expected elements.

Between steps 1 and 2, focus is lost on the text field, which triggers
the AbstractTextPropertySection.handleTextModified(), which itself
makes the actual model change. This happens in the UI thread from
inside Display.runAsyncMessages()..

The test itself runs from a different thread (WorkbenchTestable), so
in some contexts it may move to step 2 and 3 before the out-of-focus
trigger is enqueued, and the model change executed.

If this interpretation is correct, the waitAllUiEvents() inserted by
this patch after the text field has lost focus should make sure the
model change triggered in reaction is executed before we try to open
the dialog.

Change-Id: Ice3d643abf6f360bfdfb20dc62baa58b4124108b
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>