[507474] Update Guava range to [15.0,22.0)
This includes v15 which was in Neon, and v21 which will be in Oxygen,
the two Eclipse versions we support in Sirius 5.0.

Also normalize the way we depend on Guava to use Require-Bundle
everywhere. It's not the OSGi-recommended way, but it means we don't
depend on Orbit to have all the package versions right every time and
makes things simpler to manage by being consistent.

Bug: 507474
Change-Id: I2a69e5721cb90bafa58edd90f2fa50b710f87a8c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>