[515819] Do not lose the cross referencer after DND to resource root
Sometimes the model element contents and cross references are modified
before the REMOVE notification. If an object is dropped as a new
resource root, it has no container after the move to
resource.getContents(), the eContainer() is null but the element is
still contained in a resource (which has the cross ref adapter) so it
should not be removed or we will a sub-tree of the model without the
semantic cross referencer.

Bug: 515819
Cherry-picked-from: 515684
Change-Id: Ieecc9f652abf7a692745ad4fa6af6b435d5b618b
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>