[507774] Keep current feature values among valid choices
It's the job of the filters in the FeatureEditorDialog to not display
the values which are already present on the reference, but we should
still provide them so they are available to display when removed from
the right side.

Also take this opportunity to remove our own obsolete "fork" of EMF's
FeatureEditorDialog. The patch which historically justified this
forked version has long been intergrated in all versions of EMF that
we are compatible with (see bug #381535).

Bug: 507774
Change-Id: I0908382a07337cdfa2fd20dad46f0052e862a671
Signed-off-by: jmallet <jessy.mallet@obeo.fr>