[506259] Display "(transient)" for transient layers in VSM
The change is hard-coded because diagram.genmodel has not be modified(
"Edit/Style Providers" is still set to false)
If "Edit/Style Providers" was set to true, we should change all getText
and getStyledText generated methods in all xxxProvider classes of the
oes.diagram.ui plug-in. That will be done in another commit if we decide
to generalize the use of styled label.

Bug: 506259
Change-Id: I787a72dc284ae2d8559e25e7e706c1851917363a
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>