[519305] Fix GMF bug by overriding figure code in EditPart
Before this fix, a small shift can appears because of the computation
delta between using getBounds and using getHandleBounds. This is visible
with snap to grid (in some cases the result is not snapped).
All instantiations of DefaultSizeNodeFigure have been replaced by the
new SiriusDefaultSizeNodeFigure. In future, the fix will be done in GMF
(bug 519250) and Sirius will use again DefaultSizeNodeFigure.
Specific tests have been added in EdgeCreationPositionTes, but this is
in EdgeCreationPositionWithSnapToGridTest that the problem is really
visible.

Some data of EdgeWithBorderNodeCreationPositionWithSnapToGridTest have
been adapted. In fact, now the created border node is correctly snapped
to the grid, and the previous data is not on the grid.

Bug: 519305
Change-Id: I4af970d7a9658adccd91f726c03f4ddcde165757
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>