[493353] Adds updateRepresentation method on DRepresentationDescriptor
* In the case of a move representation for instance, we need to update
the repPath attribute without calling the setRepresentation method which
will send a notification. The updateRepresentation method updates the
repPath.
* This commit also modify the moveRepresentation method to update the
repPath before modifying the DRepresentationDescriptor parent. Indeed, a
notification was sent that triggered the refresh of the model explorer.
But at that time the repPath was not updated and caused trouble during
the refresh of the DRepresentationItem (descriptor.getRepresentation()
returned null).

Bug: 493353
Change-Id: I00f91d49195af0e04102f92a1c7c9c63a39c315b
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>