[510040] Removed "Remove Model" action for non external models
Removed "Remove Model" action for non external models in session
editor's models block if the project containing the session has the
nature modeling project. 

These models must not be removed from session
when they are present in the same project as the aird because they are
loaded automatically when the project has modeling nature. 

Fix checkstyle error

Bug: 510040
Change-Id: Ib1cb3bf1b72e9cb77383e49549eabdcaf324d259
Signed-off-by: pguilet <pierre.guilet@obeo.fr>