[487718] Re-introduce SubMonitor.split(int) now that Mars support is gone
Calls to SubMonitor.split(int) were introduced by e236175a1fd189d, but
at the time we still needed to build on Mars, where this API did not
exist. Commit 96c76f319efea replaced those calls with older APIs, but
it's not clear the behavior was the same. Now that Mars support is
gone, we can reintroduce the original (and more modern) code.

Bug: 487718
Change-Id: I0cc4e35ac11db16c42035f8083dc5cc01553b227
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>