[510040] Fix session editor opened twice regarding the launch context
The session editor could be launched twice when it was opened
automatically at session loading from modeling perspective when opening
a modeling project. And when opened by a double click in the java
perspective just after.

Now if the session editor is already opened, the focus is given to it in
such context.

Bug: 510040
Change-Id: I1ddf71132285e13da22fd8d90e17089804e03d91
Signed-off-by: pguilet <pierre.guilet@obeo.fr>