[516734] Adapt SiriusToolServices to use Set<> and List<>
Changing SiriusToolServices and SiriusContext to expose the actual type
of the collection being either Set or List. 
Note that getAllSemanticElements() returns a Set and
allSemanticElements() returns a List which is unexpected but matches the
actual implementation in order to avoid any change at runtime so late in
the release cycle.

Bug: 516734
Change-Id: I26036819255c103bcc6b225534589d8b015d419f
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>