[refactoring] Improve the preprocessing of the actions
The toolbar actions and the widget actions were not handled in a similar
fashion since toolbar actions were using the regular preprocessing
mechanism while the widget actions were marked as being copied and later
in PreconfiguredProcessor, the regular mechanism was executed in a
specific way for widget actions. This behavior was also partially
executed for most multi-valued features.

Change-Id: Ifce83e31955e7f4944bae9531030b54b294b8fff
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>