[510040] Fix create representation wizard permission not working
Now the user is warned if he cannot create a new representation because
he has no write access on any DView containing the representation
description from which a representation instance must be created.

If no DView contains a representation descriptor pointing at the
representation description used to create the
representation instance, then creation is allowed (this situation
happens when the viewpoint containing the description has never been
activated yet in the session).

Bug: 510040
Change-Id: Ifdfd025ad497aade46afb14ecabc066fe7872e9e
Signed-off-by: pguilet <pierre.guilet@obeo.fr>