[517437] Fix pb in case of Oblique edge with small segment
In case of an old label location outside of the segment, we ensure that
the distance between the label and the edge is not higher than before.
In this case, we reset the location to its default.
The distance is between the center of the label of the nearest segment
of the edge.

Bug: 517437
Change-Id: I6903264a06b634d877d39f794882ad8b9f5d7110
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>