[517206] Do not share content provider in different views
The content provider (and most importantly, the underlying
AdapterFactoryContentProvider) used by all instances of
SiriusCommonContentProvider was a single, shared instance. This caused
UI refresh issues when different views tried to use such providers, as
the underlying AFCP only remembers the last view it was attached to
when forwarding ViewerRefresh requests to the UI layer.

Bug: 517206
Change-Id: I422b9fed91e33d83c853c86943d66e0c02cbfc3b
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>