[517505] Add a checkbox to show disabled viewpoints in aird editor
Now the representations block of the aird editor have a checkbox "Show
disabled viewpoint".

When checked, this checkbox shows viewpoints and/or sub items
(representation description and representations) regarding status of the
checkbox "group by viewpoints" either when parent viewpoint is enabled
or disabled.

When not checked, this checkbox shows viewpoints and/or sub items
(representation description and representations) regarding status of the
checkbox "group by viewpoints" only when parent viewpoint is enabled

Bug: 517505
Change-Id: I18bbe185119975225dd2924c7a786dddb8093394
Signed-off-by: pguilet <pierre.guilet@obeo.fr>