[479049] Fix ResizeKindEditorTest
Since the commit [1], the FeatureAssistTest has been added before the
ResizeKindEditorTest in the AllTestSuite [2]. Since this time, the
ResizeKindEditorTest fails on Sirius CI server.
After analysis, the ResizeKindEditorTest is not the cause. It is the
code calls to display completion in FeatureAssistTest.
Class SiriusSWTBotTable and SiriusSWTBotTableItem have been copied from
SWTBot repository [3] master branch, on commit [4], to benefit of all
fixes for these classes. These classes must be deleted when the bugzilla
457804 [5] will be OK. They are now used instead of the ones from
current version of SWTBot.
We also use the workaround described in [6] to be sure to have an active
shell during ResizeKindEditorTest.
This commit also improves ContentAssistTest that fails on some PC.

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=02f266cd9d0d5b83346755a8edc72f66a5ff8514
[2]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/diff/plugins/org.eclipse.sirius.tests.swtbot/src/org/eclipse/sirius/tests/swtbot/suite/AllTestSuite.java?id=02f266cd9d0d5b83346755a8edc72f66a5ff8514
[3] http://git.eclipse.org/c/swtbot/org.eclipse.swtbot.git/
[4]
http://git.eclipse.org/c/swtbot/org.eclipse.swtbot.git/commit/?id=5d7ffcc5be09ba79a3ab78063b33550d655b6a9a
[5] https://bugs.eclipse.org/bugs/show_bug.cgi?id=457804
[6]
https://wiki.eclipse.org/SWTBot/Troubleshooting#No_active_Shell_when_running_SWTBot_tests_in_Xvfb

Bug: 479049
Change-Id: I5baff86f47fec68f4507521c8290b890d1ed26fb
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>