[519044] Fix bendpoints coordinates for edge creation.
Now, during edge creation, source and target connection bendpoints are
computed according to the intersection of the line relying source and
target anchor and source and target figure bound.
If edge is rectilinear, some intermediate bendpoints are added on the
edge. If the edge has three segments, two intermediate bendpoints are
added else if the edge has two segments, only one intermediate bendpoint
is added.

Bug: 519044 
Change-Id: I3dc52baf0ceea6dca76ed7bb9aa2152b45e4a6eb
Signed-off-by: jmallet <jessy.mallet@obeo.fr>