[518524] Page providing API creation
An API based on a registry and page providers has been created.
It allows to register providers of additional pages for the aird editor
with positioning information.

An ordering algorithm has been created to displays provided pages in
the right order defined by all page positioning information.

The default page now uses this API to be displayed in aird editor.

The Sirius debug plugin now provides an empty aird editor page by using
the new API when the Sirius debug view is active.

Bug: 518524
Change-Id: I60c65d6c36dacdf694e429eabd3c635a53fc0575
Signed-off-by: pguilet <pierre.guilet@obeo.fr>