[516669] Fixes editor memento when restarting Eclipse.
* When restarting eclipse with Sirius editors opened, The
DRepresentation was loaded using the URIEditorInput, without using the
DRepresentationDescriptor.getRepresentation() method. Bypass this API to
load a representation is not recommended and in that case the
DRepresentationDescriptor#representation feature is not added in the
XRef.
* The SessionEditorInput has now an additional attribute to held the
DRepresentationDescriptor URI.
* Since the SessionEditorInput expected uri value in many places is the
GMF Diagram uri, this commit does not modify the value of this
attribute. Instead, the new attribute repDescURI is used during editor
initialization to retrieve the DRepresentation by using the
DRepresentationDescriptor. This attribute is also saved and loaded in
the IMemento.
* Also fixes the SessionEditorInputTest

Bug: 516669
Change-Id: I7cfd19ad93fa96ba516faaaad0b673c1df19d743
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>