[516669] Retrieve inverse target from DRepresentationDescriptor
The DialectManager.getRepresentations(final EObject semantic, final
Session session) implementation retrieved the representations from the
DRepresentation.target inverse reference. This code have to fulfill the
contract even if representations are not loaded. In order to not load
systematically all representations, we use the
DRepresentationDescriptor.target inverse reference instead and load the
representation only for the found DRepresentationDescriptor.

Bug: 516669
Change-Id: Ia0c8086ab23ad2a09bad16dc8668b9dbcce028ce
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>