[520900] Add 4 new straighten action with a side pinned
Add 4 actions to straighten a selection of edges with a side pinned. For
a selection of horizontal edges, it would be possible to straighten them
with the left side pinned or the right side pinned. For a selection of
vertical edges, it would be possible to straighten them with the top
side pinned or the bottom side pinned.

Bug: 520900
Change-Id: I9093d794815bcbd0f2f3da30de81a049c79ec651
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>
Signed-off-by: pguilet <pierre.guilet@obeo.fr>