[520881] Complete existing tests
This commit improves the code coverage of existing tests.
It allows to find some bugs before addressing the bug 520881.
In BorderedNodeCreationTest, in assertBorderedNodeAtLocation(), we check
the Draw2D location of the border node and now the GMF location. This
required the addition of
org.eclipse.sirius.ext.gmf.runtime.editparts.GraphicalHelper.getAbsoluteBoundsIn100Percent(GraphicalEditPart,
boolean). The existing
GraphicalHelper.getAbsoluteBoundsIn100Percent(GraphicalEditPart) does
not consider the scrollbar of containers. This new API is not documented
as it is more for internal uses.

Several tests of AllCreationTestSuite (*WithSnapToGrid*) will be fixed
in following commits, for this issue or for other bugs concerning
snapToGrid (like bug 520632).

This commit also formats source code of CDBIL.

Bug: 520881
Change-Id: Ida20a610d0ab2180cb11b1af098d8f2c352c666e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>