[520881] Add snapToGrid notion in CanonicalDBorderItemLocator
This commit introduces the snapToGrid notion in
DDiagramCanonicalSynchronizer and CanonicalDBorderItemLocator. This is
the only way to have correct results (if snapToGrid is enabled). The
location stored in EdgeLayoutData was already snapped to the grid but
the CanonicalDBorderItemLocator broke this constraints as it had not
snapToGrid notion.
This commit does not fix all failing tests of previous commit but a
large part.

Bug: 520881
Change-Id: Ie61ef87284d13f535cfb037e3fe0cc51261d74ae
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>