[520881] Add sub EdgeLayoutData to handle edge creation with border node
The main goal of this commit is to add a sub EdgeLayoutData to handle
correctly the specific case of edge creation with associated border
nodes.
We only discover if data is needed for border nodes during the real GMF
View creation (DDiagramCanonicalSynchronizer). So we must store 2 data
during the tool execution and choose the right one at creation.

Bug: 520881
Change-Id: I8ea842845277684dc29bd0cf4b53a5feace987ef
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>