[518246] Fix bug when resizing execution from the top side
If the lower bound of an execution is on the limit of the lower bound of
its parent operand, the resize from the top side to top was possible
until 5 pixels. But it should be only be possible for 4 pixels.
Otherwise, there was a bug because the execution is moved above the
start of the operand.

Bug: 518246
Change-Id: I7f7c752e367b739a19cfd11911fadff85a08405f
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>