[520632] Handle other cases where Sirius default PADDING is used
This commit is not really in scope of the issue but when
SiriusLayoutDataManager.PADDING is used, if snapToGrid is enabled we
must always consider the grid spacing instead of the default PADDING.
Even if code concerned in AirXYLayoutEditPolicy is not used, it has been
changed too.
The changed code for SiriusLayoutDataManagerImpl is tested with new test
in GroupElementsInOneOtherTests.

Bug: 520632
Change-Id: I73dcfe8eab5529d81c58063a753306e47e22f0d9
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>