[521467] Restore binary compatibility for DialectUIServices.export() methods
Commit 4dad490670711724edb changed the signature of the export()
methods to return an ExportResult instead of void. This is
source-compatible but breaks binary compatibility with client code
built against previous versions, causing java.lang.NoSuchMethodError
when it tries to export a representation.

To keep binary compatiblity, restore the original methods with their
signatures, which simply call *new* methods named exportWithResult().
The new methods return the ExportResult, and the old "void export()"
simply call exportWithResult() and ignore the return value.

Bug: 521467
Change-Id: I34b2a392985f1418c5141e3e7c8efe64a4add890
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>