[525299] Fixes wrong border node location with snap to grid
* If the grid spacing was equal to the border node size, the first
border node was located on the grid but the CanonicalDBorderItemLocator
located the next border node with 1 px gap out of the grid. This gap is
the minimal distance between two border nodes.
* Now if the next location with the 1px gap is greater than the next
grid step, we try to located the border node on the following grid step.
* The tests from PortLocationTest were not written to be compliant with
snapToGrid so we disable it to avoid unexpected problem.

Bug: 525299
Change-Id: I77c7d2e63070caa91cd4fcb312ac2c0e9f034d42
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>