[506259] Fix export of diagram with decorations
 * To be exported in the diagram image, the decorations have to be in
DiagramRootEditPart.DECORATION_PRINTABLE_LAYER. Consequently before
creating the diagram edit part used to create the image, we set
SiriusDiagramUiPreferencesKeys.PREF_PRINT_DECORATION preference to true.
The preference is reseted just after export to its previous value.
 * Note that the process may fail if, during the time between the moment
we start creating the diagramPart and the moment the
SiriusGenericDecorator reads the preference value, the value of the
preference is changed by another thread. The case is not significant and
moreover a third party code has no reason to change that preference.

Bug: 506259
Change-Id: I32a0d9e107f6de927e961b4b491e44cb404318f4
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>