[527047] Capability to adapt size when exporting diagrams as image
Before this issue, there was only the capability to activate the
auto-scaling. With this option, generated image was bigger, had best
quality and was longer to generate. Now there is the capability to adapt
between nominal size (as with previous auto-scaling option disabled)
and maximum size of image (as with previous auto-scaling option
enabled).

Bug: 527047
Change-Id: Ic709adf0cfefdd307b3d8702bd1a8fae91923ef3
Signed-off-by: jmallet <jessy.mallet@obeo.fr>