[527109] update behavior when showing mode is active
This mode triggered by a button in the tabbar now shows all invisible
elements as well as visible ones without changing their visibility
status from a persistence model point of view.

Invisible elements are shown with transparency. This feature as done for
a POC is not working for all parts.

Bug: 527109
Change-Id: Iec17d9dc63dbbded2c17c503af3da4929b545fd6
Signed-off-by: pguilet <pierre.guilet@obeo.fr>