[527676] Add test for double-click tool dynamic availability
When defining a double click tool on a VSM, it is now immediately
available on diagram in the same workspace using this VSM.

The actual fix was made in 5f0b59fa8b6d88e0c862bd676c1ae19c378768f6 as
a side-effect (probably unintended) of work for bug #527109, which
modifies DoubleClickEditPolicyProvider.java. This simply adds the
corresponding non-regression test.

Bug: 527676
Change-Id: I2bca354aed67affb51db4351d50d25db8db836a1
Signed-off-by: pguilet <pierre.guilet@obeo.fr>