[521718] Fix listener not cleared from diagram event broker after reload
The removal of the listeners associated to DView in the diagram broker
is now done when the session is about to be replaced in addition of when
it has already been replaced.

It allows to find the DView associated to the
listeners to remove because the listener's semantic element is still
attach to the AirdResource. Whereas it is not the case when the session
has been replaced. That event only occurs in
SessionResourcesSynchronizer.reloadResource().

Then we are able to remove the entry
identified by both the DView and its listener.



Bug: 521718
Change-Id: Ie11e964f449f477bb0517d77c0a2fccbeae78ea0
Signed-off-by: pguilet <pierre.guilet@obeo.fr>