[527109] Change double click behavior while in showing mode
A double click done on a diagram edit part now switch its visibility
status instead of applying any existing double click tool.

As a reminder, the goal of the four commits concerning 527109 is a POC.
Several tasks are remaining to have a fully operational feature. A
complete list will be added directly in bug 527109, a not exhaustive
list is:

-Double click should change the hidden status and not the visibility one
that is caused by several reasons.
-Double click on an hidden edge should also reveal source and target
node
-Check double click policy on all kind of parts
-The showing mode button's icon could be improved
-Layouting and showing mode should be exclusive.
-Transparency on hidden elements should be available for all parts.

Bug: 527109
Change-Id: I248779b3f44a20ca5e8b600a406d0fe33aba0bc4
Signed-off-by: pguilet <pierre.guilet@obeo.fr>