[525933] Fix regression in test test_DnDClassFromMC2DiagramBlank5
For commit [1] 641c44b9, an existing representation description of a VSM
has been completed with a transient layer. But this new transient layer
has a side effect on an existing test. This commit adds a new
representation description and a new corresponding representation for
the test test_DnDClassFromMC2DiagramBlank6FromATransientLayerTool(). The
test test_DnDClassFromMC2DiagramBlank5() is no longer impacted.

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=641c44b9275f78ce9ed31b72fabe0bdf44ff100e

bug: 525933
Change-Id: I33ba3a50d438a4f7f133939e509f91844d591b67
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>