[521467] Increase delta used when testing for exported image size
On some systems, exported images are 2 pixels wider than expected:

   expected:<4956.0> but was:<4958.0> 

It's not clear why exactly, but it may be related to different fonts
or DPIs. It does not seem worth failing the tests, especially given
that the image export API does not provide any strong guarantee on the
actual sizes.

Bug: 521467
Change-Id: I0f754173a27a6412dc91520f391cdf452d4c90d4
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>