[529148] Add action to insert vertical blank space in sequence diagram
With this commit, it is now possible to insert vertical blank space in
sequence diagram by using a drag'n'drop in the ruler grid with Ctrl key
pressed.

In future, to allow insertion of horizontal blank space in sequence
diagram and, vertical and horizontal blank space in standard diagram,
the methods
SequenceEMFCommandFactory.buildInsertHorizontalBlankSpaceCommand(DDiagram,
int,
int)
UndoRedoCapableEMFCommandFactory.buildInsertVerticalBlankSpaceCommand(DDiagram,
int, int),
UndoRedoCapableEMFCommandFactory.buildInsertHorizontalBlankSpaceCommand(DDiagram,
int, int) will have to be implemented.

This commit uses a new SiriusRulerComposite to have a specific
RulerEditPartFactory to instantiate a specific SiriusRulerEditPart to
finally enable a specific SimpleDragTracker when Ctrl key is pressed.
This SimpleDragTracker, SiriusRulerDragTracker, draws the feedback and
executes the command returned by the SiriusRulerEditPart.
The SiriusRulerEditPart delegates the creation of the command to
IDiagramCommandFactory.
Currently, only insertion of vertical blank space in sequence diagram is
implemented.

Bug: 529148
Change-Id: I06e6a45eb24e93f955c8aa28c715ef0b869e9068
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>