[542859] Extract coercion rules into IConverters
Separate the actual evaluation of the expressions, done by the
evaluate(EObject, String) method, from the conversion into the result
type expected by the multiple evaluateXXX() methods. This makes the
evaluateXXX() methods of all leaf interpreters (except the
DefaultInterpreter, which is a rather special case) trivial and same
across all implementations, something like:

  return converter.convert(evaluate(target, expr)).toXXX();

Most actual interpreters use the same coercion rules now defined in
the new DefaultConverter class, which are now explicit and can be
unit-tested (not done in this patch). The two execptions are OCL and
Acceleo/MTL for which the existing rules are kept thanks to custom
IConverter implementations, even though they do not exactly match what
all the other interpreters do.

This is a first step towards replacing all the evaluateXXX() methods
with a single, cleaner and more powerful evaluate() method coupled
with a rich evaluation result type.

Bug: 542859
Change-Id: Icd43c17264f83222c2ee3b12982a766d6b79860d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>