[534301] POC of Popup Menu and actions improvement
To enable this POC, you must add the VM arg
"-Dorg.eclipse.sirius.experimental.popupMenuImprovement=true".

This POC can be tested with the project [1]
org.eclipse.sirius.sample.popupmenu attached to bug 529992.

This POC allows to:
* Add group/separator to the root of the contextual menu (Popup Menu "/"
in sample)
* Add popup menu in a specific group (Popup Menu "POC subMenu2" in
sample)
* Add action in a specific group (Operation Action "POC action31", "POC
action32", "POC action33" in sample)
* Add group/separator in an existing menu (Popup Menu "Edit" in sample)
* Add action in an existing menu (Operation Action "POC action01", "POC
action02", "POC action03", "POC action04" in sample)

The result is visible in image POC_contextualMenu.png [2] and the same
menu without the POC is visible in contextualMenu_withoutPOC.png [3].

This POC does not allow to Add a group/separator before or after an
existing one. The new groups are automatically added at the end.

[1] https://bugs.eclipse.org/bugs/attachment.cgi?id=273596
[2] https://bugs.eclipse.org/bugs/attachment.cgi?id=273595
[3] https://bugs.eclipse.org/bugs/attachment.cgi?id=273638

Bug: 534301
Cherry-pick-of: 529992
Change-Id: I5c31d374b1e77d1d8d5f9a3305d7fe6815e25419
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>