[test] Add new API in tests framework to load a odesign file
The method SiriusTestCase.loadModeler(URI, EditingDomain) allows to load
the VSM at the specified URI and register all its Viewpoints in the
testcase. There is now a new method, SiriusTestCase.loadModeler(URI,
EditingDomain, boolean), to allow to load the VSM at the specified URI
without registering its viewpoints. It can be useful to load a VSM for
an editing domain of another session than the current session of the
testcase.

Change-Id: I39789e714421f29a4d699518448c8164b15f4ccf
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>