[530372] Fix viewpoint activation side effects
Changing the activation status of a viewpoint in an session editor now
does not change the activation status of this viewpoint in other opened
session editors.

This patch changes the equal method of ViewpointItemImpl and
RepresentationDescriptionItemImpl to discriminate items from one session
to items from another and avoid mixing those in different aird editor.

Bug: 530372
Change-Id: I1faa95854e197886f8db3b2aac9622a92b7dfca6
Signed-off-by: pguilet <pierre.guilet@obeo.fr>