[529879] Fix viewpoint activation side effects
Changing the activation status of a viewpoint in an session editor now
does not change the activation status of this viewpoint in other opened
session editors.

The label provider TooltipAwareNavigatorDecoratingLabelProvider has been
removed because it causes the problem and it is not necessary to display
tooltips in viewpoints viewer of session editor whereas it is the only
reason for its usage.

Bug: 529879
Change-Id: I19eefde42f77242812143f4adaa62444c03d7f32
Signed-off-by: pguilet <pierre.guilet@obeo.fr>