[509070] ELK algorithms integration to Sirius
ELK algorithms have been integrated to Sirius.

A specifier can now add in the VSM a GenericLayout pointing at the id of
an ELK algorithm. This algorithm will then be used to do the arrange-all
action.

A registry has been made allowing to provide generic layout providers to
be use from the VSM directly.

This registry can accept any provider subclassing DefaultLayoutProvider.
Not only the ELK provider.

ELK integration is now available with Sirius build.

An ELK feature has been created and is present in the Sirius update site
under the Sirius category.

Bug: 509070
Change-Id: Ifdd28184fa2b1c6d5a118b212f5d9e9f15093ce5
Signed-off-by: pguilet <pierre.guilet@obeo.fr>